New Activities Added
Addition of activities for seeing available caravans, creating a new
caravan, and selecting which of these two paths to follow.  Also Changed
the manifest such that the vehicle details entry page is the new launch
point.  The layouts for these activities contain the required UI
elements for the most part but the arrangement of those parts has not
been nailed down.