Added Class & Adapter for AvailableCaravans Activity
Created an "Available" class to store all of the information about an
available caravan for usage with the adapter that was written to
populate the list of caravans.  Currently the phone can request the list
of caravans from the V2X, get it back, and populate the array.  However,
the elements in the array do not appear in the listview of the UI yet.