Shuffling of Function Calls
Relocated the setAdapter call which was causing the listview of
available caravans not to be populated.  Corrected wrong message values
for sending join_caravan message over BT to the V2X.

Next steps: Implement behavior for when a join is granted or denied,
create the "lobby" screen for the Host which is the mirror of
AvailableCaravans for the members.