Another Day
Created page for the host to see who is in the caravan.  Added a kick
button, will make that work someday but it's not very high on the
agenda.  Also created the (blank) main activity where all the exciting
stuff will take place once I have creating and joining sorted out.
Popups now appear letting the members know if their request to join was
succesful and if not, why.