Fragments, so many fragments
Support for using spaces in caravan and vehicle names (this needs to be adjusted for vehicle make and model as well in a future update).
Created main activity with fragments for all of the various activities which may be undertaken.  Most work so far has been done on the messaging activitiy thus far.
Data validation added to pages with user input prompts.
Started rearranging layouts of vehicle and caravan details pages so they make more sense to users, still pretty ugly though, need to talk to John for suggestions.